sc: remove redundant casts
surprisingly GetDouble() already returns a double

+++ sc/source/core/tool/interpr4.cxx
@@ -2729,7 +2729,7 @@ void ScInterpreter::ScExternal()
                     break;
 
                 case SC_ADDINARG_DOUBLE:
-                    aParam <<= (double) GetDouble();
+                    aParam <<= GetDouble();
                     break;
 
                 case SC_ADDINARG_STRING:
@@ -2841,7 +2841,7 @@ void ScInterpreter::ScExternal()
                             {
                                 uno::Any aElem;
                                 if ( nStackType == svDouble )
-                                    aElem <<= (double) GetDouble();
+                                    aElem <<= GetDouble();
                                 else if ( nStackType == svString )
                                     aElem <<= GetString().getString();
                                 else
@@ -2887,7 +2887,7 @@ void ScInterpreter::ScExternal()
                     switch( nStackType )
                     {
                         case svDouble:
-                            aParam <<= (double) GetDouble();
+                            aParam <<= GetDouble();
                             break;
                         case svString:
                             aParam <<= GetString().getString();
